Improve error handling in apt setup script
`wget -qO- ... | sudo tee` causes any errors from `wget` to be swallowed up, causing an empty keyring file to be configured.

For us, the local Ubuntu setup did not have `ca-certificates` installed, which caused `wget` to error out because of a authentication errors (untrusted certificate). However, the `-q` flag and the pipe into `tee` prevented the error from showing up or halting the script.

This PR turns on non-verbose output for `wget` and adds a proper short-circuit on the return code of it by storing the keyring in a temporary file first.